Fold new tests into already existing cases. Add turbulent forcing test case. Not added FlameSheet turbulent forcing as I do not think its testing anything additional than HIT.
