+++ Exec/RegTests/HIT/README
@@ -1,0 +1,32 @@
+## HIT
+Two different turbulent options are possible (HIT decay and maintained HIT).
+
+## HITDecay
+This is the decaying 3D Homogeneous Isotropic Turbulence test case,
+originally implemented in PeleC and used to study compressible turbulence in:
+E. Motheau and J. Wakefield, "Investigation of finite-volume methods to capture shocks and turbulence spectra in compressible flows", Commun. in Appl. Math. and Comput. Sci, 15-1 (2020), 1--36.
+https://arxiv.org/abs/1902.06665
+
+The script gen_hit_ic.py from PelePhysics/Support/TurbFileHIT can be used to generate an initial solution.
+
+Here, there is no scaling by the turbulent Mach number and the density is kept constant, so that the Taylor-scale Reynolds number is defined by Re=0.5 urms / nu where umrs can be set in the input file (prob.urms0=...) and nu is computed from the mixture composition.
+
+At the beginning of a simulation, a file initialConditions.txt is created to hold the values of urms0, lambda0 and tau, which will be used for post-processing.
+
+## TurbForce Incompressible
+A periodic domain that can be used to generate and maintain HIT. This is typically used to generate the inital turbulence profile used in maintained HIT flame sheets (see Exec/RegTests/TurbForce_FlameSheet).
+
+# Citation
+To cite the forcing scheme, please cite the [CAMCOS article](http://dx.doi.org/10.2140/camcos.2008.3.103)
+```
+@article{aspden2008analysis,
+  title={{Analysis of implicit LES methods}},
+  author={Aspden, Andrew and Nikiforakis, Nikos and Dalziel, Stuart and Bell, John},
+  journal={Communications in Applied Mathematics and Computational Science},
+  volume={3},
+  number={1},
+  pages={103--126},
+  year={2008},
+  publisher={Mathematical Sciences Publishers}
+}
+```
