+++ Exec/RegTests/HIT/README.md
@@ -1,5 +1,11 @@
-## TurbForce Incompressible
+## HIT
+Two different turbulent options are possible (HIT decay and maintained HIT).
 
+## HITDecay
+A 3D decaying Homogeneous Isotropic Turbulence (HIT) case, where the initial solution is generated with a Passot-Pouquet spectrum. Testing the basic incompressible
+integration algorithm and Large Eddy Simulation implementation.
+
+## TurbForce Incompressible
 A periodic domain that can be used to generate and maintain HIT. This is typically used to generate the inital turbulence profile used in maintained HIT flame sheets (see Exec/RegTests/TurbForce_FlameSheet).
 
 # Citation
