+++ Exec/RegTests/HIT/hit_decay_ic_4_32.dat
