+++ Exec/RegTests/HIT/input.HITDecay.3d
@@ -11,7 +11,7 @@ peleLM.lo_bc = Interior  Interior  Interior
 peleLM.hi_bc = Interior  Interior  Interior
 
 #-------------------------AMR CONTROL----------------------------
-amr.n_cell          = 32 32 32      # Level 0 number of cells in each direction   
+amr.n_cell          = 32 32 32      # Level 0 number of cells in each direction
 amr.v               = 1                # AMR verbose
 amr.max_level       = 0                # maximum level number allowed
 amr.ref_ratio       = 2 2 2 2          # refinement ratio
@@ -24,7 +24,7 @@ amr.max_grid_size   = 64               # max box size
 #----------------------- PROBLEM PARAMETERS---------------------
 prob.T_mean           = 298.0
 prob.P_mean           = 101325.
-prob.input_name       = hit_ic_4_32.dat
+prob.input_name       = hit_decay_ic_4_32.dat
 prob.input_resolution = 32
 prob.input_binaryformat = 0
 prob.urms0            = 2.0
@@ -49,7 +49,7 @@ amr.derive_plot_vars = avg_pressure mag_vort
 # Refinement according to the vorticity, no field_name needed
 amr.refinement_indicators     = magvort
 amr.magvort.max_level         = 3
-amr.magvort.vorticity_greater = 500 
+amr.magvort.vorticity_greater = 500
 
 #--------------------------OMP TILE INPUTS-----------------------------
 #fabarray.mfiter_tile_size = 8 8 8
