+++ Exec/RegTests/HIT/input.HITDecay.3d_les
@@ -14,22 +14,33 @@ peleLM.hi_bc = Interior  Interior  Interior
 amr.n_cell          = 32 32 32      # Level 0 number of cells in each direction
 amr.v               = 1                # AMR verbose
 amr.max_level       = 0                # maximum level number allowed
-# amr.ref_ratio       = 2 2 2 2          # refinement ratio
-# amr.regrid_int      = 2                # how often to regrid
+amr.ref_ratio       = 2 2 2 2          # refinement ratio
+amr.regrid_int      = 2                # how often to regrid
 amr.n_error_buf     = 2 2 2 2          # number of buffer cells in error est
 amr.grid_eff        = 0.7              # what constitutes an efficient grid
 amr.blocking_factor = 16               # block factor in grid generation (min box size)
 amr.max_grid_size   = 64               # max box size
 
 #----------------------- PROBLEM PARAMETERS---------------------
-prob.input_name       = hit_ic_4_32.dat
+prob.T_mean           = 298.0
+prob.P_mean           = 101325.
+prob.input_name       = hit_decay_ic_4_32.dat
 prob.input_resolution = 32
 prob.input_binaryformat = 0
 prob.urms0            = 2.0
 
+#-----------------------DIFFUSION AND LES MODEL CONTROL-----------------------
+peleLM.Prandtl = 0.7                   # [OPT, DEF=0.7] If fixed_Pr or doing LES, specifies the Prandtl number
+peleLM.Schmidt = 0.7                   # [OPT, DEF=0.7] If doing LES, specifies the Schmidt number
+peleLM.les_model = "Smagorinsky"              # [OPT, DEF="None"] Model to compute turbulent viscosity: None, Smagorinsky, WALE
+peleLM.les_cs_smag = 0.18              # [OPT, DEF=0.18] If using Smagorinsky LES model, provides model coefficient
+peleLM.les_cm_wale = 0.60              # [OPT, DEF=0.60] If using WALE LES model, provides model coefficient
+peleLM.les_v = 1                       # [OPT, DEF=0] Verbosity level for LES model
+peleLM.plot_les = 1
+
 #-------------------------PeleLM CONTROL----------------------------
 peleLM.v = 3
-peleLM.incompressible = 1
+peleLM.incompressible = 0
 peleLM.rho = 1.17
 peleLM.mu = 0.00001
 peleLM.num_init_iter = 3
