+++ Exec/RegTests/HIT/input.HITForced.3d
@@ -1,0 +1,49 @@
+#---------------------- DOMAIN DEFINITION ------------------------
+geometry.is_periodic = 1 1 1                 # For each dir, 0: non-perio, 1: periodic
+geometry.coord_sys   = 0                     # 0 => cart, 1 => RZ
+geometry.prob_lo     = 0.0    0.0    0.0     # x_lo y_lo (z_lo)
+geometry.prob_hi     = 0.06283185307179586232 0.06283185307179586232 0.06283185307179586232     # x_hi y_hi (z_hi)
+
+#---------------------- BC FLAGS ---------------------------------
+# Interior, Inflow, Outflow, Symmetry,
+# SlipWallAdiab, NoSlipWallAdiab, SlipWallIsotherm, NoSlipWallIsotherm
+peleLM.lo_bc = Interior Interior Interior           # bc in x_lo y_lo (z_lo)
+peleLM.hi_bc = Interior Interior Interior           # bc in x_hi y_hi (z_hi)
+
+#---------------------- AMR CONTROL ------------------------------
+amr.n_cell          = 32 32 32           # Level 0 number of cells in each direction
+amr.v               = 1                  # AMR verbose
+amr.max_level       = 0                  # maximum level number allowed
+amr.ref_ratio       = 2 2 2 2            # refinement ratio
+amr.regrid_int      = 2                  # how often to regrid
+amr.n_error_buf     = 2 2 2 2            # number of buffer cells in error est
+amr.grid_eff        = 0.7                # what constitutes an efficient grid
+amr.blocking_factor = 16                 # block factor in grid generation (min box size)
+amr.max_grid_size   = 64                 # max box size
+
+#---------------------- PeleLMeX CONTROL -------------------------
+peleLM.v = 3                              # PeleLMeX verbose
+peleLM.incompressible = 1
+peleLM.mu   = 1.e-5
+peleLM.rho  = 1.03
+
+#---------------------- Time Stepping CONTROL --------------------
+amr.max_step = 100                         # Maximum number of time steps
+amr.cfl = 0.6                              # CFL number for hyperbolic system
+amr.dt_shrink = 1.0                        # Scale back initial timestep
+amr.stop_time = 1.0
+
+#---------------------- IO CONTROL -------------------------------
+amr.check_int = 100                       # Frequency of checkpoint output
+amr.plot_int = 50                         # Frequency of pltfile output
+amr.derive_plot_vars = turbforces
+
+#---------------------- Linear solver CONTROL --------------------
+mac_proj.verbose = 0
+nodal_proj.verbose = 0
+
+#---------------------- Turbulent Forcing CONTROL-------------------------
+peleLM.do_turbulent_forcing = 1
+turbforce.rho_incompressible = 1.03
+turbforce.urms = 1.5
+turbforce.v = 1
