--- Exec/RegTests/HITDecay/README
@@ -1,10 +1,0 @@
-This is the decaying 3D Homogeneous Isotropic Turbulence test case,
-originally implemented in PeleC and used to study compressible turbulence in:
-E. Motheau and J. Wakefield, "Investigation of finite-volume methods to capture shocks and turbulence spectra in compressible flows", Commun. in Appl. Math. and Comput. Sci, 15-1 (2020), 1--36.
-https://arxiv.org/abs/1902.06665
-
-The script gen_hit_ic.py from PelePhysics/Support/TurbFileHIT can be used to generate an initial solution.
-
-Here, there is no scaling by the turbulent Mach number and the density is kept constant, so that the Taylor-scale Reynolds number is defined by Re=0.5 urms / nu where umrs can be set in the input file (prob.urms0=...) and nu is computed from the mixture composition.
-
-At the beginning of a simulation, a file initialConditions.txt is created to hold the values of urms0, lambda0 and tau, which will be used for post-processing.
