--- Exec/RegTests/TurbForce_Incompressible/CMakeLists.txt
@@ -1,8 +1,0 @@
-set(PELE_PHYSICS_EOS_MODEL Fuego)
-set(PELE_PHYSICS_CHEMISTRY_MODEL air)
-set(PELE_PHYSICS_TRANSPORT_MODEL Simple)
-set(PELE_PHYSICS_ENABLE_SPRAY OFF)
-set(PELE_PHYSICS_SPRAY_FUEL_NUM 0)
-set(PELE_PHYSICS_ENABLE_SOOT OFF)
-set(PELE_PHYSICS_ENABLE_RADIATION OFF)
-include(BuildExeAndLib)
