--- Exec/RegTests/TurbForce_Incompressible/pelelmex_prob.cpp
@@ -1,14 +1,0 @@
-#include <PeleLMeX.H>
-#include <AMReX_ParmParse.H>
-
-void
-PeleLM::readProbParm()
-{
-  amrex::ParmParse pp("turbforce");
-  pp.query("urms", PeleLM::prob_parm->urms);
-}
-
-void
-PeleLM::freeProbParm()
-{
-}
