+++ Tests/CMakeLists.txt
@@ -243,8 +243,9 @@ if(NOT PELE_ENABLE_EB)
     add_test_rt(isothermal-soret IsothermalSoretChannel)
   endif()
   if(PELE_DIM EQUAL 3)
-    add_test_r(hit-${PELE_DIM}d HITDecay)
-    add_test_r(hit-les-${PELE_DIM}d HITDecay)
+    add_test_r(hit-decay-${PELE_DIM}d HIT)
+    add_test_r(hit-decay-les-${PELE_DIM}d HIT)
+    add_test_r(hit-forced-${PELE_DIM}d HIT)
   endif()
 else()
   if(PELE_DIM EQUAL 2)
